flux-sample-app-2/cd_flux/requirements.txt
boto3==1.9.252
botocore==1.12.252
aws-xray-sdk==2.4.2
gitpython